Address the rest of Shams comments
