+++ gradle-plugins/cordformation/src/main/kotlin/net/corda/plugins/Cordformation.kt
@@ -32,6 +32,7 @@ class Cordformation : Plugin<Project> {
     override fun apply(project: Project) {
         Utils.createCompileConfiguration("cordapp", project)
         // Add runtime dependency to node except for the corda project itself to avoid a circular dependency of Corda with itself.
+        // It's used to pregenerate network parameters on deployNodes task.
         if (project.rootProject.name != "corda-project") {
             project.dependencies.add("runtime", "net.corda:corda-node:${project.rootProject.ext<String>("cordaVersion")}")
         } else {
