+++ gradle-plugins/cordformation/src/main/kotlin/net/corda/plugins/Node.kt
@@ -3,9 +3,7 @@ package net.corda.plugins
 import com.typesafe.config.*
 import net.corda.cordform.CordformNode
 import org.gradle.api.Project
-import sun.security.x509.X500Name
 import java.io.File
-import java.io.IOException
 import java.nio.charset.StandardCharsets
 import java.nio.file.Files
 import java.nio.file.Path
@@ -121,12 +119,11 @@ class Node(private val project: Project) : CordformNode() {
             project.logger.error("Node has a null name - cannot create node")
             throw IllegalStateException("Node has a null name - cannot create node")
         }
-
         val dirName = try {
-            // Using sun.security.x509.X500Name because importing BouncyCastle provider in Cordformation causes problems
+            // Parsing O= part directly because importing BouncyCastle provider in Cordformation causes problems
             // with loading our custom X509EdDSAEngine.
-            X500Name(name).organization
-        } catch(_ : IOException) {
+            name.trim().split(",").filter { it.startsWith("O=") }.first().substringAfter("=")
+        } catch(_ : NoSuchElementException) {
             // Can't parse as an X500 name or no organisation part, use the full string
             name
         }
